Moved user input functions to one file
